Refactor Emitters split multiple files
