+++ src/compiler/emitters/wirespec/src/commonMain/kotlin/community/flock/wirespec/emitters/wirespec/WirespecEndpointDefinitionEmitter.kt
@@ -1,0 +1,34 @@
+package community.flock.wirespec.emitters.wirespec
+
+import community.flock.wirespec.compiler.core.emit.EndpointDefinitionEmitter
+import community.flock.wirespec.compiler.core.emit.EndpointEmitter
+import community.flock.wirespec.compiler.core.emit.IdentifierEmitter
+import community.flock.wirespec.compiler.core.emit.Spacer
+import community.flock.wirespec.compiler.core.emit.TypeDefinitionEmitter
+import community.flock.wirespec.compiler.core.parse.Endpoint
+import community.flock.wirespec.compiler.core.parse.Field
+
+interface WirespecEndpointDefinitionEmitter: EndpointEmitter, TypeDefinitionEmitter, EndpointDefinitionEmitter, IdentifierEmitter {
+
+    override fun emit(endpoint: Endpoint) = """
+        |endpoint ${emit(endpoint.identifier)} ${endpoint.method}${endpoint.requests.emitRequest()} ${endpoint.path.emitPath()}${endpoint.queries.emitQuery()} -> {
+        |${endpoint.responses.joinToString("\n") { "$Spacer${it.status.fixStatus()} -> ${it.content?.reference?.emit() ?: "Unit"}${if (it.content?.reference?.isNullable == true) "?" else ""}" }}
+        |}
+        |
+    """.trimMargin()
+
+    private fun List<Endpoint.Segment>.emitPath() = "/" + joinToString("/") {
+        when (it) {
+            is Endpoint.Segment.Param -> "{${it.identifier.value}: ${it.reference.emit()}}"
+            is Endpoint.Segment.Literal -> it.value
+        }
+    }
+
+    private fun List<Endpoint.Request>.emitRequest() =
+        firstOrNull()?.content?.reference?.emit()?.let { " $it" }.orEmpty()
+
+    private fun List<Field>.emitQuery() = takeIf { it.isNotEmpty() }
+        ?.joinToString(",", "{", "}") { it.emit() }
+        ?.let { " ? $it" }
+        .orEmpty()
+}
